refactor: move enums to entities package to improve module organization and clarity
