+++ src/main/java/com/nexaworks/rafiq/dto/request/medicine/MedicineFilter.java
@@ -1,7 +1,7 @@
 package com.nexaworks.rafiq.dto.request.medicine;
 
-import com.nexaworks.rafiq.enums.MedicineStatus;
-import com.nexaworks.rafiq.enums.MedicineType;
+import com.nexaworks.rafiq.entities.enums.MedicineStatus;
+import com.nexaworks.rafiq.entities.enums.MedicineType;
 
 public record MedicineFilter(String search, MedicineStatus status, String groupId,
         MedicineType type) {
