+++ src/main/java/com/nexaworks/rafiq/dto/request/medicine/UpdateMedicinePatchRequest.java
@@ -3,9 +3,9 @@ package com.nexaworks.rafiq.dto.request.medicine;
 import java.time.Instant;
 import java.util.Optional;
 
-import com.nexaworks.rafiq.enums.MedicineFrequency;
-import com.nexaworks.rafiq.enums.MedicineStatus;
-import com.nexaworks.rafiq.enums.MedicineType;
+import com.nexaworks.rafiq.entities.enums.MedicineFrequency;
+import com.nexaworks.rafiq.entities.enums.MedicineStatus;
+import com.nexaworks.rafiq.entities.enums.MedicineType;
 
 public record UpdateMedicinePatchRequest(Optional<String> name, Optional<String> dosage,
         Optional<String> notes, Optional<MedicineFrequency> frequency, Optional<Instant> startDate,
