+++ src/main/java/com/nexaworks/rafiq/dto/request/medicine/UpdateMedicineRequest.java
@@ -2,9 +2,9 @@ package com.nexaworks.rafiq.dto.request.medicine;
 
 import java.time.Instant;
 
-import com.nexaworks.rafiq.enums.MedicineFrequency;
-import com.nexaworks.rafiq.enums.MedicineStatus;
-import com.nexaworks.rafiq.enums.MedicineType;
+import com.nexaworks.rafiq.entities.enums.MedicineFrequency;
+import com.nexaworks.rafiq.entities.enums.MedicineStatus;
+import com.nexaworks.rafiq.entities.enums.MedicineType;
 
 import jakarta.validation.constraints.NotNull;
