+++ src/main/java/com/nexaworks/rafiq/dto/response/Group/AddGroupResponse.java
@@ -2,7 +2,7 @@ package com.nexaworks.rafiq.dto.response.Group;
 
 import java.util.UUID;
 
-import com.nexaworks.rafiq.enums.Color;
+import com.nexaworks.rafiq.entities.enums.Color;
 
 public record AddGroupResponse(UUID groupId, UUID patientId, String description, Color color,
         String name, Long createdAt, Long updatedAt, String iconUrl, int medicineCount) {
