+++ src/main/java/com/nexaworks/rafiq/dto/response/Group/GroupDetailsResponse.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.UUID;
 
 import com.nexaworks.rafiq.dto.response.medicine.MedicineGroupResponse;
-import com.nexaworks.rafiq.enums.Color;
+import com.nexaworks.rafiq.entities.enums.Color;
 
 public record GroupDetailsResponse(UUID id, UUID userId, String name, String description,
         Color color, String iconUrl, int medicineCount, List<MedicineGroupResponse> medicines,
