+++ src/main/java/com/nexaworks/rafiq/dto/response/Group/GroupResponse.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.UUID;
 
 import com.nexaworks.rafiq.dto.response.medicine.MedicinePreview;
-import com.nexaworks.rafiq.enums.Color;
+import com.nexaworks.rafiq.entities.enums.Color;
 
 public record GroupResponse(UUID id, String name, String description, Color color, String iconUrl,
         int medicineCount, List<MedicinePreview> medicinePreviews) {
