+++ src/main/java/com/nexaworks/rafiq/dto/response/medicine/MedicineGroupResponse.java
@@ -2,8 +2,8 @@ package com.nexaworks.rafiq.dto.response.medicine;
 
 import java.util.UUID;
 
-import com.nexaworks.rafiq.enums.MedicineFrequency;
-import com.nexaworks.rafiq.enums.MedicineStatus;
+import com.nexaworks.rafiq.entities.enums.MedicineFrequency;
+import com.nexaworks.rafiq.entities.enums.MedicineStatus;
 
 public record MedicineGroupResponse(UUID id, String name, String dosage,
         MedicineFrequency frequency, MedicineStatus Status) {
