+++ src/main/java/com/nexaworks/rafiq/dto/response/reminder/AddReminderResponse.java
@@ -4,9 +4,9 @@ import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
 
-import com.nexaworks.rafiq.enums.Day;
-import com.nexaworks.rafiq.enums.MedicineStatus;
-import com.nexaworks.rafiq.enums.ReminderFrequency;
+import com.nexaworks.rafiq.entities.enums.Day;
+import com.nexaworks.rafiq.entities.enums.MedicineStatus;
+import com.nexaworks.rafiq.entities.enums.ReminderFrequency;
 
 public record AddReminderResponse(UUID id, UUID userId, UUID medicineId, String medicineName,
         String dosage, String time, ReminderFrequency frequency, List<Day> dayOfWeek,
