+++ src/main/java/com/nexaworks/rafiq/dto/response/reminder/ReminderResponse.java
@@ -2,8 +2,8 @@ package com.nexaworks.rafiq.dto.response.reminder;
 
 import java.util.UUID;
 
-import com.nexaworks.rafiq.enums.ReminderFrequency;
-import com.nexaworks.rafiq.enums.ReminderStatus;
+import com.nexaworks.rafiq.entities.enums.ReminderFrequency;
+import com.nexaworks.rafiq.entities.enums.ReminderStatus;
 
 public record ReminderResponse(UUID id, String time, ReminderFrequency frequency,
         ReminderStatus status) {
