+++ src/main/java/com/nexaworks/rafiq/entities/Medicine.java
@@ -4,9 +4,9 @@ import java.time.Instant;
 import java.util.HashSet;
 import java.util.Set;
 
-import com.nexaworks.rafiq.enums.MedicineFrequency;
-import com.nexaworks.rafiq.enums.MedicineStatus;
-import com.nexaworks.rafiq.enums.MedicineType;
+import com.nexaworks.rafiq.entities.enums.MedicineFrequency;
+import com.nexaworks.rafiq.entities.enums.MedicineStatus;
+import com.nexaworks.rafiq.entities.enums.MedicineType;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
