+++ src/main/java/com/nexaworks/rafiq/entities/User.java
@@ -8,7 +8,7 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.nexaworks.rafiq.enums.Gender;
+import com.nexaworks.rafiq.entities.enums.Gender;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Email;
