+++ src/main/java/com/nexaworks/rafiq/entities/enums/Color.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.enums;
+package com.nexaworks.rafiq.entities.enums;
 
 public enum Color {
     RED, GREEN, BLUE, YELLOW, BLACK, WHITE, ORANGE, VIOLET, PINK, BROWN, SILVER, GREY, MAROON, TEAL, NAVY, OLIVE, PURPLE
