+++ src/main/java/com/nexaworks/rafiq/entities/enums/Gender.java
@@ -1,0 +1,5 @@
+package com.nexaworks.rafiq.entities.enums;
+
+public enum Gender {
+    MALE, FEMALE
+}
