+++ src/main/java/com/nexaworks/rafiq/entities/enums/MedicineStatus.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.enums;
+package com.nexaworks.rafiq.entities.enums;
 
 public enum MedicineStatus {
     ACTIVE, DISCONTINUED, INACTIVE
