+++ src/main/java/com/nexaworks/rafiq/entities/enums/MedicineType.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.enums;
+package com.nexaworks.rafiq.entities.enums;
 
 public enum MedicineType {
     TABLET, CAPSULE, LIQUID, INJECTION, TOPICAL, INHALER, SUPPOSITORY, PATCH, DROPS, POWDER, PRESCRIPTION, SUPPLEMENT, OTHER
