+++ src/main/java/com/nexaworks/rafiq/entities/enums/ReminderFrequency.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.enums;
+package com.nexaworks.rafiq.entities.enums;
 
 public enum ReminderFrequency {
     DAILY, WEEKLY, MONTHLY, YEARLY, CUSTOM
