+++ src/main/java/com/nexaworks/rafiq/entities/enums/ReminderStatus.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.enums;
+package com.nexaworks.rafiq.entities.enums;
 
 public enum ReminderStatus {
     UPCOMING, TAKEN, MISSED, SNOOZED, SKIPPED
