+++ src/main/java/com/nexaworks/rafiq/entities/enums/Roles.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.enums;
+package com.nexaworks.rafiq.entities.enums;
 
 public enum Roles {
     ROLE_USER, ROLE_ADMIN, ROLE_DOCTOR, ROLE_PATIENT
