+++ src/main/java/com/nexaworks/rafiq/entities/enums/Status.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.enums;
+package com.nexaworks.rafiq.entities.enums;
 
 public enum Status {
     IN_REVIEW, VERIFIED, REJECTED
