+++ src/main/java/com/nexaworks/rafiq/entities/enums/TokenType.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.enums;
+package com.nexaworks.rafiq.entities.enums;
 
 public enum TokenType {
     OTP, REFRESH, ACCESS_TOKEN, JWT_BLACKLIST
