--- src/main/java/com/nexaworks/rafiq/enums/Gender.java
@@ -1,5 +1,0 @@
-package com.nexaworks.rafiq.enums;
-
-public enum Gender {
-    MALE, FEMALE
-}
