+++ src/main/java/com/nexaworks/rafiq/service/ImageService.java
@@ -5,7 +5,7 @@ import java.io.IOException;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.nexaworks.rafiq.dto.client.cloundinary.UploadResults;
-import com.nexaworks.rafiq.enums.UploadType;
+import com.nexaworks.rafiq.entities.enums.UploadType;
 
 public interface ImageService {
     UploadResults uploadResource(MultipartFile file, UploadType type) throws IOException;
