+++ src/main/java/com/nexaworks/rafiq/service/RoleService.java
@@ -1,7 +1,7 @@
 package com.nexaworks.rafiq.service;
 
 import com.nexaworks.rafiq.entities.Role;
-import com.nexaworks.rafiq.enums.Roles;
+import com.nexaworks.rafiq.entities.enums.Roles;
 
 public interface RoleService {
     Role getRole(Roles roles);
