+++ src/main/java/com/nexaworks/rafiq/service/ServiceImpl/DoctorServiceImpl.java
@@ -8,7 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.nexaworks.rafiq.dto.client.cloundinary.UploadResults;
 import com.nexaworks.rafiq.entities.DoctorProfile;
 import com.nexaworks.rafiq.entities.User;
-import com.nexaworks.rafiq.enums.Status;
+import com.nexaworks.rafiq.entities.enums.Status;
 import com.nexaworks.rafiq.exception.custom.UserNotFoundException;
 import com.nexaworks.rafiq.repository.DoctorRepository;
 import com.nexaworks.rafiq.service.DoctorService;
