+++ src/main/java/com/nexaworks/rafiq/service/ServiceImpl/TokenServiceImpl.java
@@ -10,7 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.nexaworks.rafiq.entities.Token;
 import com.nexaworks.rafiq.entities.User;
-import com.nexaworks.rafiq.enums.TokenType;
+import com.nexaworks.rafiq.entities.enums.TokenType;
 import com.nexaworks.rafiq.exception.custom.TokenInvalidException;
 import com.nexaworks.rafiq.exception.custom.TokenNotFoundException;
 import com.nexaworks.rafiq.exception.custom.UserException;
