+++ src/main/java/com/nexaworks/rafiq/service/ServiceImpl/UserServiceImpl.java
@@ -1,7 +1,5 @@
 package com.nexaworks.rafiq.service.ServiceImpl;
 
-import static com.nexaworks.rafiq.enums.Roles.*;
-
 import java.io.IOException;
 import java.time.Instant;
 import java.util.Optional;
@@ -25,7 +23,7 @@ import com.nexaworks.rafiq.entities.PatientProfile;
 import com.nexaworks.rafiq.entities.Role;
 import com.nexaworks.rafiq.entities.Token;
 import com.nexaworks.rafiq.entities.User;
-import com.nexaworks.rafiq.enums.TokenType;
+import com.nexaworks.rafiq.entities.enums.TokenType;
 import com.nexaworks.rafiq.exception.custom.InvalidPasswordException;
 import com.nexaworks.rafiq.exception.custom.RegistrationException;
 import com.nexaworks.rafiq.repository.UserRepository;
