+++ src/test/java/com/nexaworks/rafiq/integration/controller/AuthControllerIntegrationTest.java
@@ -28,8 +28,8 @@ import com.nexaworks.rafiq.dto.request.user.VerifyOtpRequest;
 import com.nexaworks.rafiq.entities.Role;
 import com.nexaworks.rafiq.entities.Token;
 import com.nexaworks.rafiq.entities.User;
-import com.nexaworks.rafiq.enums.Gender;
-import com.nexaworks.rafiq.enums.TokenType;
+import com.nexaworks.rafiq.entities.enums.Gender;
+import com.nexaworks.rafiq.entities.enums.TokenType;
 import com.nexaworks.rafiq.integration.BaseIntegrationTest;
 import com.nexaworks.rafiq.repository.RoleRepository;
 import com.nexaworks.rafiq.repository.TokenRepository;
