+++ src/test/java/com/nexaworks/rafiq/integration/controller/GroupControllerIntegrationTest.java
@@ -27,9 +27,9 @@ import com.nexaworks.rafiq.entities.Medicine;
 import com.nexaworks.rafiq.entities.PatientProfile;
 import com.nexaworks.rafiq.entities.Role;
 import com.nexaworks.rafiq.entities.User;
-import com.nexaworks.rafiq.enums.Color;
-import com.nexaworks.rafiq.enums.Gender;
-import com.nexaworks.rafiq.enums.MedicineFrequency;
+import com.nexaworks.rafiq.entities.enums.Color;
+import com.nexaworks.rafiq.entities.enums.Gender;
+import com.nexaworks.rafiq.entities.enums.MedicineFrequency;
 import com.nexaworks.rafiq.integration.BaseIntegrationTest;
 import com.nexaworks.rafiq.repository.DrugRepository;
 import com.nexaworks.rafiq.repository.GroupRepository;
