+++ src/test/java/com/nexaworks/rafiq/unit/service/CloudinaryServiceTest.java
@@ -18,7 +18,7 @@ import org.springframework.web.multipart.MultipartFile;
 import com.cloudinary.Cloudinary;
 import com.cloudinary.Uploader;
 import com.nexaworks.rafiq.dto.client.cloundinary.UploadResults;
-import com.nexaworks.rafiq.enums.UploadType;
+import com.nexaworks.rafiq.entities.enums.UploadType;
 import com.nexaworks.rafiq.exception.custom.EmptyFileException;
 import com.nexaworks.rafiq.exception.custom.FileException;
 import com.nexaworks.rafiq.exception.custom.FileUploadException;
