+++ src/test/java/com/nexaworks/rafiq/unit/service/MedicineServiceImplTest.java
@@ -21,8 +21,8 @@ import com.nexaworks.rafiq.entities.Drug;
 import com.nexaworks.rafiq.entities.Group;
 import com.nexaworks.rafiq.entities.Medicine;
 import com.nexaworks.rafiq.entities.PatientProfile;
-import com.nexaworks.rafiq.enums.Action;
-import com.nexaworks.rafiq.enums.MedicineStatus;
+import com.nexaworks.rafiq.entities.enums.Action;
+import com.nexaworks.rafiq.entities.enums.MedicineStatus;
 import com.nexaworks.rafiq.exception.custom.GroupNotFoundException;
 import com.nexaworks.rafiq.exception.custom.MedicineAlreadyExist;
 import com.nexaworks.rafiq.exception.custom.MedicineLimit;
