+++ src/test/java/com/nexaworks/rafiq/unit/service/TokenServiceTest.java
@@ -15,7 +15,7 @@ import org.springframework.test.util.ReflectionTestUtils;
 
 import com.nexaworks.rafiq.entities.Token;
 import com.nexaworks.rafiq.entities.User;
-import com.nexaworks.rafiq.enums.TokenType;
+import com.nexaworks.rafiq.entities.enums.TokenType;
 import com.nexaworks.rafiq.exception.custom.TokenInvalidException;
 import com.nexaworks.rafiq.exception.custom.TokenNotFoundException;
 import com.nexaworks.rafiq.exception.custom.UserException;
